requirements.txt
streamlit>=1.32.0
google-generativeai>=0.8.0
plotly>=5.18.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
pydantic>=2.5.0
numpy>=1.26.0
pandas>=2.1.0
